+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for G733Z LED modes
+- Support for GU604V LED modes
 ### Changed
 - Adjustments to Anime system events thread
 - Add "sleep" animetion config options to anime config